Slightly improve the data model

+++ src/main/java/org/students/simplebitcoinnode/repository/MerkleTreeNodeRepository.java
@@ -3,6 +3,6 @@ package org.students.simplebitcoinnode.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.students.simplebitcoinnode.entity.MerkleTreeNode;
 
-public interface MerkleTreeNodeRepository extends JpaRepository<MerkleTreeNode, Integer> {
+public interface MerkleTreeNodeRepository extends JpaRepository<MerkleTreeNode, Long> {
 
 }
